refactor(tasks): share the streaming line splitter between file tasks
`linesFromStream` was a private generator inside `FileGrepTask.server`.
`FileSedTask.server` needs the same cap and must not carry a second copy of
it, so it moves to `util/StreamLines.server` unchanged.

No behavior change: the grep path imports the same generator it defined.
